Remove EjectMethod
