+++ main.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"strings"
 
 	"github.com/bitrise-io/go-utils/command"
 	"github.com/bitrise-io/go-utils/errorutil"
@@ -24,93 +23,6 @@ type Config struct {
 	Password                   stepconf.Secret `env:"password"`
 	RunPublish                 string          `env:"run_publish"`
 	OverrideReactNativeVersion string          `env:"override_react_native_version"`
-}
-
-// EjectMethod if the project is using Expo SDK and you choose the "bare" --eject-method those imports will stop working.
-type EjectMethod string
-
-// const ...
-const (
-	Bare    EjectMethod = "bare"
-	ExpoKit EjectMethod = "expoKit"
-)
-
-// Expo CLI
-type Expo struct {
-	Version string
-	Method  EjectMethod
-	Workdir string
-}
-
-// installExpoCLI runs the install npm command to install the expo-cli
-func (e Expo) installExpoCLI() error {
-	args := []string{"install", "-g"}
-	if e.Version != "latest" {
-		args = append(args, "expo-cli@"+e.Version)
-	} else {
-		args = append(args, "expo-cli")
-	}
-
-	cmd := command.New("npm", args...)
-	cmd.SetStdout(os.Stdout)
-	cmd.SetStderr(os.Stderr)
-
-	log.Donef("$ " + cmd.PrintableCommandArgs())
-	return cmd.Run()
-}
-
-// Login with your Expo account
-func (e Expo) login(userName string, password stepconf.Secret) error {
-	args := []string{"login", "--non-interactive", "-u", userName, "-p", string(password)}
-
-	cmd := command.New("expo", args...)
-	cmd.SetStdout(os.Stdout)
-	cmd.SetStderr(os.Stderr)
-
-	nonFilteredArgs := ("$ " + cmd.PrintableCommandArgs())
-	fileredArgs := strings.Replace(nonFilteredArgs, string(password), "[REDACTED]", -1)
-	log.Printf(fileredArgs)
-
-	return cmd.Run()
-}
-
-// Logout from your Expo account
-func (e Expo) logout() error {
-	cmd := command.New("expo", "logout", "--non-interactive")
-	cmd.SetStdout(os.Stdout)
-	cmd.SetStderr(os.Stderr)
-
-	log.Printf("$ " + cmd.PrintableCommandArgs())
-	return cmd.Run()
-}
-
-// Eject command creates Xcode and Android Studio projects for your app.
-func (e Expo) eject() error {
-	args := []string{"eject", "--non-interactive"}
-
-	cmd := command.New("expo", args...)
-	cmd.SetStdout(os.Stdout)
-	cmd.SetStderr(os.Stderr)
-	if e.Workdir != "" {
-		cmd.SetDir(e.Workdir)
-	}
-
-	log.Donef("$ " + cmd.PrintableCommandArgs())
-	return cmd.Run()
-}
-
-func (e Expo) publish() error {
-	args := []string{"publish", "--non-interactive"}
-
-	cmd := command.New("expo", args...)
-	cmd.SetStdout(os.Stdout)
-	cmd.SetStderr(os.Stderr)
-	if e.Workdir != "" {
-		cmd.SetDir(e.Workdir)
-	}
-
-	log.Donef("$ " + cmd.PrintableCommandArgs())
-	return cmd.Run()
 }
 
 func parsePackageJSON(pth string) (serialized.Object, error) {
@@ -167,23 +79,8 @@ func main() {
 		failf("Input validation error: %s", err)
 	}
 
-	//
-	// Select the --eject-method
-	ejectMethod := Bare
-	fmt.Println()
-	log.Infof("Define --eject-method")
-	{
-		if cfg.UserName != "" {
-			ejectMethod = ExpoKit
-			log.Printf("Expo account credentials have provided => Set the --eject-method to %s", ejectMethod)
-		} else {
-			log.Printf("Expo account credentials have not provided => Set the --eject-method to %s", ejectMethod)
-		}
-	}
-
 	e := Expo{
 		Version: cfg.ExpoCLIVersion,
-		Method:  ejectMethod,
 		Workdir: cfg.Workdir,
 	}
 
@@ -196,39 +93,6 @@ func main() {
 			failf("Failed to install the selected (%s) version for Expo CLI, error: %s", cfg.ExpoCLIVersion, err)
 		}
 	}
-
-	//
-	// Logging in the user to the Expo account
-	fmt.Println()
-	log.Infof("Login to Expo")
-	{
-		switch ejectMethod {
-		case ExpoKit:
-			if err := e.login(cfg.UserName, cfg.Password); err != nil {
-				failf("Failed to log in to your provided Expo account, error: %s", err)
-			}
-		case Bare:
-			log.Printf("--eject-method has been set to bare => Skip...")
-		}
-	}
-
-	//
-	// Logging out the user from the Expo account at the end of the step (even if it fails)
-	defer func() {
-		fmt.Println()
-		log.Infof("Logging out from Expo")
-		{
-			if e.Method == ExpoKit {
-				if err := e.logout(); err != nil {
-					log.Warnf("Failed to log out from your Expo account, error: %s", err)
-				}
-			} else if e.Method == ExpoKit {
-				log.Printf("Logout input was set to false => Skip...")
-			} else {
-				log.Printf("You were not logged in => Skip...")
-			}
-		}
-	}()
 
 	//
 	// Eject project via the Expo CLI
@@ -245,12 +109,7 @@ func main() {
 	log.Donef("Successfully ejected your project")
 
 	if cfg.RunPublish == "yes" {
-		fmt.Println()
-		log.Infof("Running expo publish")
-
-		if err := e.publish(); err != nil {
-			failf("Failed to publish project: %s", err)
-		}
+		runPublish(e, cfg)
 	}
 
 	if cfg.OverrideReactNativeVersion != "" {
@@ -296,5 +155,37 @@ func main() {
 			}
 			failf("%s failed: %s", cmd.PrintableCommandArgs(), err)
 		}
+	}
+}
+
+func runPublish(expo Expo, cfg Config) {
+	//
+	// Logging in the user to the Expo account
+	fmt.Println()
+	log.Infof("Login to Expo")
+	{
+		if err := expo.login(cfg.UserName, cfg.Password); err != nil {
+			failf("Failed to log in to your provided Expo account, error: %s", err)
+		}
+	}
+
+	//
+	// Logging out the user from the Expo account (even if it fails)
+	defer func() {
+		fmt.Println()
+		log.Infof("Logging out from Expo")
+		{
+			if err := expo.logout(); err != nil {
+				log.Warnf("Failed to log out from your Expo account, error: %s", err)
+			}
+		}
+	}()
+
+	fmt.Println()
+	log.Infof("Running expo publish")
+
+	// Running publish
+	if err := expo.publish(); err != nil {
+		failf("Failed to publish project: %s", err)
 	}
 }
